SUTT_3/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-allauth==0.50.0
django-filter==21.1
django-import-export==2.8.0
django-sendgrid-v5==1.2.0
et-xmlfile==1.1.0
google-api-python-client==1.6.4
gunicorn==20.1.0
heroku==0.1.4
httplib2==0.20.4
idna==3.3
jsonpickle==0.9.6
MarkupPy==1.14
more-itertools==8.13.0
oauth2client==4.1.2
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==1.5
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sendgrid==6.9.7
six==1.16.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
starkbank-ecdsa==2.0.3
tablib==3.2.1
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
uritemplate==3.0.1
urllib3==1.26.9
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
